requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
fastapi==0.120.4
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
